Use _TOMLDecoder as single value decoder
This makes code more shared between keyed and unkeyed containers.
It might be slower slightly but we'll worry about that later.
