+++ .swiftformat
@@ -1,3 +1,3 @@
---exclude **/*.Generated.swift
+--exclude Tests/**/*.Generated.swift
 --extension-acl on-declarations
 --ifdef no-indent
